Add function to generate computer's random choice
In order to make the choices all equally likely, I declared the
choiceNumber variable with a value of 0 then set up a while loop that
runs while choiceNumber === 0 so that a random number between 1 and
99 is generated. If 0 had been included, the number of possible random
numbers would not have been divisible by 3.

An if/else statement then turns the number into rock, paper or scissors.